The main addition in this commit is that the GM now changes maps over 3 ticks so that components and such can get all the messages before things actually happen.  All of the code for changing maps has been moved to an external class - MapChangeStateData.  The GM could easily be extended to allow a developer to subclass the MapChangeStateData class.  The name of the class is probably not the best, but it works for the moment.  Create and Delete messages are not sent during map change, as was the case before the change, but no unit test exists to verify that. It should be added.  A problem with the unit tests exists where it crashes intermittently in the sigslot preframe message passing code. This is probably due to something being deleted during the frame that is listening to preframe and ends up being next in the list in sigslot.
The scons build has been modified to allow turning off building of the examples and demos so that incremental builds don't have
to check all the dependencies all the time.  The precompiled header option name has been changed to "pch" rather than "usePrefixHeaders" 
since that terminology seems to more common among build scripts I've seen.  A test build mode has been added called "fastBuild" that
generates .cxx files for each build that include all of the .cpp files specified and compiles them together.  This option cuts the build time
in half or better for full builds, and it's actually not so bad for incremental builds since it speeds up the link times quite a bit.
The only real issue is that dtCore takes forever and certain problems had to be hacked around since some files can't be built with others.

All of the other changes are fixes where there were classes or structs named the same thing in different cpp files, which caused problems with the 
"fastBuild".  This really shouldn't be done anyway since the compiler can get confused about which one to use at runtime.
There were a couple other cleanups that were needed for the fastBuild the work, but nothing that wasn't reasonable.