removed physics update flag from transformable and made it use the collisionGeometry method to figure out if it needs to run the updates Also added a Matrix accessor for the matrix node.  The idea is that the user would go through that method so that we can use it to invalidate any caching we do on transforms in the future.
Fixed a crash in the objectruntimemappinginfo object that would occur if the user passed in a unique id that is actually the
same pointer as one in one of the maps.

Transformable actor proxy now uses relative positions for the translation, rotation, and scale properties because this makes more sense
when we add heirarchy to stage.  Plus it's a lot faster.

Added some extra exception handling to the hla component handle when an attribute or parameter name specified in the configuration
is not in the fom.