fixed a bit of formatting in the viewport and viewport manager classes.  Removed some "" includes and replaced them with <> in camera.h in stage.
Mainly, I added support for converting the RTIObjectIDStruct into a string and back
in the rprparameter translator.  I need this for time sync since the id passed is
not for an actor that will be mapped into delta3d.  That is, I just need the string name
to compare against.