Added outgoing articulation parameter translation to the RPRParameterTranslator.
Added a unit test for outgoing articulation translation.

Added IsEqual functions to articulation related objects for easier object comparison at the highest level object, ArticulatedParameter. IsEqual was used rather than an operator override so that it is apparent in the object interface that it has the capability to be compared with another of its own kind. Special IsEqual has special functionality on ParameterValue objects.

Added referencial accessors to ParameterValue for accessing AttachedParts and ArticulatedParts.

Fixed the Tag::Decode function and made a slight optimization to it.