Added the sequence support to the schema and to the loader.  It is not completely tested, but it's definitely not more broken that it was before adding the code.  Added unit tests to load the animatables, and the base properties are unit tested.  A few bugs were found and squashed in the process.  Still need to test that channels and sequences are getting their properties passed correctly.  Sequences are not currently looking up their child channels at load time, so they still have no children.
added a line to the main test runner to setup delta data automatically because several tests couldn't be run alone due to not setting it.  Now no test should have to do that.  The code was removed from proxy tests to verify.

dtAnim tests had not been added to the build in SCons.  That has been fixed.