SCONS:
fixed some issues with paths in the scons files allowing the prefix and rti path to be relative on the command line, fixed a minor macosx issue.  Also fixed the building of testAnim

dtAnim:

Clone on animatable now takes the Cal3DModelWrapper as an argument so that it can be set on the animatables being cloned.  The helper now uses this.  Also added a unit test for it and for animation sequences. 

dtGame:

Removed the Nulling of data members in the GM from the shutdown.  These objects are created in constructor and don't hold onto anything that could cause shutdown problems and most won't be deleted by setting them to null away. Additionally, they leave the GM in a bad state that would cause it to crash if any methods were called.  The GM is destroyed right after shutdown is called anyway.  The unit tests were crashing because of this nulling behavior, so this has been fixed.