Fixed several icky game manager bugs.
IsInGM is not set to false when a Remote actor is deleted.
Setting both remote and published would result in a actor being in the GM in a corrupted stated.
Throwing an exception in OnEnteredWorld would leave the GM in a bad state when it happened in both AddActor or during
a map change.

Also reworked a long clunky method: ClearTimer.  The code works the same way but now is much shorted and readable.

exception had some broken, zig-zagged comment formatting.  This has been fixed.

Unit test were written for all the fixes.