Updated shader code. Fixed 2 bugs: a) clone in shader.cpp was setting the vertex and fragment file names backwards and b) the Clear() method did not unassign pre-existing shaders, which didn't clear parameters and such.
Also, added a new method, ReloadAndReassignShaderDefinitions() to ShaderManager. If you call this method from a test key or something, the artist can completely reload all the shader files and uniforms at runtime (ie, test without restarting the app). This is in relation to the Shader article for Game Programming Gems 7. Also, this method still needs a unit test.