Fixed a potentially serious crash bug that could be caused by deleting a local actor that is published on a network.
The crash was caused by an invalid object handle pointer being passed to the RTI ambassador's "deleteObjectInstance" function. The pointer became invalid right after removing the object mappings.

Now the handle value is captured, the object mappings are removed, and then the RTI removes the object instance; the order of operation has been preserved since it seems the object deletion should be the absolute last step. A try-catch has been placed around the RTI function just for the slim case that the funtion may throw an exception.