This is a big change.  The Game Manager now supports many maps, not just one map.  The old "ChangeMap" exists, but it calls "ChangeMapSet" which takes a vector of map names.  There were several maps message classes that all had the same basic parameter, a map name, have all be merged into one class called "MapMessage" that has a group property and an accessor method to get/set a list of map names.
The logger binary format now saves as version 1.1 and has a list of maps in the key frames.  All of the messages and classes for the logger have been updated to also except for the message that returns a list of keyframes since it uses the deprecated message parameter list thing.  This should be updated to used a group and things will be much easier.  The logger should still open 1.0 logfiles without a problem.

StringUtils now has a function for generating a padded string from a number so it can be used for generating sortable indices in a group parameter.  This is handy and used for the map messages now.  Anyone using a group parameter like a vector could benefit from this.

NamedGroupParameter now has a ForEach method that takes a unary function that will be called for each child parameter of the group.  This keeps one from having to copy the list of parameters before iterating over it.

Tests were added to test all the above code.  Tests were also added to check that game events are added and removed properly when changing maps.  It also checks to see that shutdown works and that the maps are unloaded and cleared on shutdown.