fixed a couple g++ compile errors by adding typename in some templates.
Added a code fix to a couple sconscripts so they don't error out when some libraries are missing.

The Cal3DDatabase now has a method to wipe the database.  The unit tests use it.  

The hardware skinning now uses the shader manager, but there is no way yet to specify which
shader group to use.  It has a default group and name, but the character XML needs to be 
modified to allow setting this.  The name of the bone transform uniform is also hard coded.

The hardware skinned actors will now share VBO's if they use the same model.