This commit contains a few changes, but they can't be committed separately.
ActorProxy.h now has a templated GetProperty method that works like the CreateActor and CreateMessage method on the GM.
The Game manager no longer delivers messages to actors that are deleted but haven't been removed.
The Game manager Preframe method has been refactored into a number of separate protected methods.  This makes the code a bunch clearer.  The code could still stand to have some of the loops changed to for_eachs and such.