Refactored dtCore::Transform.  It now holds a matrix as a data member, and never decomposes the scale.  It has methods to change the scale, such as Rescale and CalcScale, but they do full polar decomps to adjust it.  Seting the rotation wipes the scale.  Methods for getting and setting quaternions  and getting and setting specific rows were added as well.
Delta has been updated to reflect these changes.

GetTranslation and GetRotation that return by reference were removed because they can't be implemented the way the internal data was changed.  They were added for performance reasons before, but now that all the data isn't being copied several times, it's really not a big deal.  We could add them to return by copy if necessary.  

Some build system stuff was updated in an attempt to try to make g++ builds us the export macros.  This would be handy so that building with it would yield the same errors as leaving an export macro on windows.  

dtCore::Object now has a model transform so that it can be rotated, scaled, and transformed, all using dtCore::Transform.  It has a separate accessor for setting the scale that will actually be applied on an UpdateCallback so that it will never do more than one polar decomp a frame, and the properties, which will be added later, won't be doing them.