Removed and extra command processing block that was obviously redundant and wrong. The bad code looks to be a result of a copy-and-paste oversight.
NOTE: There seems to be no unit tests for the audio manager except for initialization. Extra protected methods will have to be added to the manager to access the sound object to test effects of sound commands. Currently time limitations and lack of established unit tests for the audio manager prohibit a unit test for this code change. Other applications on our side seem to work correctly after this change.