Added the ref string.   It shares copies of strings so that frequently duplicated strings, like property names, are all pointing to the same block of memory.  The class does it's best to behave like a string as much as possible via overloaded operators.
All of the strings are stored in a hash_set, which performs a lot better than an std::set.  vc++ and g++ both have implementations of this.  It would fail to an std::set otherwise.  

In debug with a hash_set, the code runs somewhat slower than just creating new strings.  Optimized builds run somewhat faster than creating new strings.  Using an std::set seems take twice as long that a hash_set to lookup strings, at least in debug.  The memory savings can be on the order of 10s of megabytes.

Many classes have been changes because data members are now ref strings and some constructors take them.  I changed all of the constants to be ref strings because copying a refstring doesn't require a lookup, which makes creating actor properties and message parameters much faster.  

A #define exists in refstring.cpp to disable to string table and just create new strings all the time.