Build breaker (potentially).
Fixed crash in the AnimNodeBuilder that could happen if the create function was called at the wrong time; mostly in OSG 2.6.0.  AnimNodeBuilder will now defer the creation of the actual geometry until the animated character is being rendered in the scene.  This should ensure that a valid OpenGL context is valid for the AnimNodeBuilder to use. SF #2058095