Added code to verify that an ActorProxy has both a non-null actor and a valid actor type.  It should not get past CreateActorProxy with an null actor or actor type now.  Also, it checks in the getters and gives the user a long explanation of what the problem is if they create an actor with "new".
While I was hacking up the header anyway, I changed the init a little. ActorProxy now has an Init method that is virtual that calls SetActorType, CreateActor, and BuildPropertyMap.  GameActor overrides init to call BuildInvokables.  This means that the actor creation routine stuff isn't spread around several classes, and it means that uses can override init if they want to add something to the creation routine.

This will help with the code in SimCore where we keep having to initialize the helper in CreateActor.

The one odd side effect is that BuildInvokables is now called before SetGameManager, since that is called by the GM still after the creation.  This is how it worked not too long ago before we set the gm on creation.  Hopefully no one should be relying on the GM existing in build invokables anyway since you can't really register for messages despite the fact that the actor has a point to the GM.