* Split Ground Clmaper into Base Ground Clamper and Default Ground Clamper. * Primary Ground Clamper implementation has been kept in Default Ground Clamper. * Default Ground Clamper has had smaller methods added to reduce the size of the larger methods and make the clamper easier to unit test and sub-class. * Several unit tests have been added for most of the changes.
NOTE: These classes may still be refactored in the future with further design discussions. The original Ground Clamper did not make it easy to implement custom behavior for other terrains that could not be detected with Isectors.

Current changes give a little more control but still are not as straight forward as one would hope. If terrains allowed for common interface, ground clamping optimized code would be much more cleaner and legible.