Stopped using QSound in favor of OpenAL sound.  QSound supports a few more sound codecs than OpenAL, and we don't want to give users the wrong impression when they preview sounds from STAGE (I had compressed .wav files playing in STAGE that won't play in a Delta application).
This required direct use of OpenAL as there is no way to play sounds from dtAudio when you're not inside a game/application loop.  dtAudio (or possibly the upcoming dtQt) should probably be modified in the future so as to support playing sounds without having to fire up an application.