Updated FileUtils::IsSameFile method to (1) not depend on path names to determines a file uniqueness, and (2) work without depending on inodes (which Windows doesn't support).
Added a unit test that ensures FileUtils::FileCopy will not blow away a file when an attempt is made to copy it onto itself (the whole reason for the creation of FileUtils::IsSameFile was to avoid this).

Updated FileUtils::DirCopy so that it won't attempt to overwrite directories if the overwrite flag is set to false.