Added support for ODE regular cylinders (as opposed to the "capped" cylinders which we used to only support).  Regular cylinders work much better for modeling motion of some wheeled vehicles.  (user submitted)
User's notes:
"I noticed Delta3d only supports ODE's CappedCylinders.  Those really don't work so good for a wheelchair because it adds hemispheres to the faces of the wheel cylinder making it difficult to fit through doors etc.  I have noticed a question that never seemed to get fully resolved in the forums regarding this, but nothing since 2007.  ODE originally supported only CappedCylinders, but added support for normal ones years ago.

"I also fixed where it renders the collision geometry.  Previously even though the geometry was a capped cylinder, it still rendered as a normal cylinder which is very misleading and is what caused me all this trouble other than blindly thinking CCylinders were the same as Cylinders :-)  It renders a OSG::Capsule which I believe to be the same geometry as ODE's capped cylinder.

"The capped cylinders were previously listed in places as 'CYLINDER' I added 'CCYLINDER' and made the new cylinders use the previous "CYLINDER" and the old capped ones use 'CCYLINDER'"