AudioManager now stores the OpenAL device and context that it's using.  The device is also retrievable via a Getter.
Cleanup was also added to the destructor for context and device.

This was done to solve some problems when recording and listening OpenAL simultaneously  (with playing taken care of via dtAudio and recording done custom -- recording not yet supported by dtAudio).  If a machine has multiple sound devices (sound cards) then you probably want to ensure that recording and playback are handled by the same card.