Second commit for trying to clean up windows.h includes.  (sorry, forgot a comment on the first commit)
These add an mswin.h to dtUtil that wrapps up the windows include, and also tried to clean up some of the undef stuff, and tried to only include windows.h when necessary.  The theory is that delta compiles so much faster on linux and osx partly because windows.h was being included pretty much everywhere.