Fixed the shutdown on the audio manager. alutExit must be called when the context is still valid.  It will also not delete contexts it didn't create.
If you pass in a context/device pair to Instantiate, then the audio manager can be flagged to destroy or not destroy them on shutdown.  It still defaults to shutting them down since that was the old behavior.

In the past, the audio manager was essentially never shutting down alut.

The IsInitialized method is now static because, with it not being static, you could only really call it if was initialized.  Essentially the method was theoretically useless when it wasn't static.  Practically, however, it worked because the method didn't access "this" so it didn't crash even though it was called on a null object.

Added some more unit tests to test the corrected behavior.