refactored the AIUtility's AIComponent to derive from dtAI::BaseAIComponent (including removing redundant code).
Moved the auto-adding of the AIPluginInterface's debug drawable out of BaseAIComponent::ProcessMessage()'s INFO_MAP_LOADED message handling and into the corresponding place for AIUtility's AIComponent. This way the AIUtility will automatically get this debug drawable, but anything else that uses or derives from the BaseAIComponent isn't forced to manually remove it if they don't want it.