Made the sound actor proxy properties still hold their values even if it can't allocate a source in the audio manager.  This will prevent map corruption in stage.
Also added a bunch more unit tests for it, and fixed the timer testing unit tests.  They have been commented out because they failed intermittently.  There were numerous bad assumptions made in the code, include the fact that IsPlaying() == true doesn't work because IsPlaying() returns an int and the value is never 1.  It also takes the elapsed time directly from the system so that it uses the EXACT elapsed time that the GM used rather than using a timer and hoping they match.