Improved execution speed of the (non-hierarchical) AStar implementation by changing the container used for the closed list from std::vector to std::set. Much time had been wasted doing linear searches on the closed list vector to see if node data was on it, so now these searches are done in logarithmic time on the closed list set.
Internal consequences: Since I was abandoning std::find_if for greener pastures and the std::set doesn't have a find() method that allows for a functor to be used to retrieve data, I had to change the closed list container to contain the data itself rather than the allocated nodes containing the data. Consequently the memory allocation (mainly for freeing) scheme had to be rearranged somewhat, so now all allocated nodes are managed on a single mDeleteMe vector that is added to whenever nodes are created and that is freed and cleared whenever FreeMem() is called.

External consequences: because std::set::find() involves searching a binary search tree, the given node_type::data_type must now implement the < operator.