Added vsync and multi-sample to the application config interface, that is you can put them into the window element in the config.xml file.
This made me have to break the CreateInstances interface, so I made it more future proof by passing the full app config, not just a few parameters of it.  Since this method can't practically be overridden, and it's called internally, I decided it probably won't break any existing code that doesn't already not work properyl, such as overriding that method or calling it externally.

At the same time, I realized that the window traits were not being passed to the osgadaperwidgets in the QFormat.  This meaning adding a new constructor and changing the glwidgetfactory interface.  I did this in a way that should be backward compatible, and I deprecated the old method.