The property macros now work for olmost all property types include enumeration, game event, group (actor property, but not get/set), ActorID, and resource.  ActorActor probably will never be supported, and Resources only support the new mode of just getting/setting a resource descriptor.
To make an Enumeration work for the get/set, you have to declare the type as dtUtil::EnumerationPointer<EnumSubClass>, which is a holder type that will allow setting the pointer from a reference.  

Creating a Vec[2,3,4] get/set will end up creating a Vec[2,3,4]fActorProperty because of the typedef.  As a result, I changed Vec[2,3,4]ActorProperty to just be typedefs of the "f" version, and the datatypes VEC[2,3,4] are now just references to the "F" versions.  I had to add support for overriding the Enum to string comparison so that I could make, for example, GetValueForName("VEC3") return the VEC3F type.

Finally, I noticed that the Enumeration creation macros were inlining all of the functions, so I changed it so the implementations for the two big ones now go in the cpp.  AddInstance() is only called at startup and GetValueForName() is a loop over a vector with string compares, so it's unlikely to every be inline anyway.  This should cut down the preprocessed header sizes by 100s of lines of code because we declare lots of enums.

I would have made these a separate commit, but it's all connected.  I fixed a bug in Map where if you remove an actor and have an ActorIDActorProperty that points to it, it wouldn't clear it out properly, NOT ONLY that, but it was also static_cast-ing it to an ActorActorProperty and calling SetValue(NULL), which was causing intermittent crashes in the unit tests.  There was a similar bug in the actual unit tests I also fixed.  The tests don't crash randomly for me any longer.