Major refactor of the way the actor components are initialized in Delta. I implemented this according to what Erik and David had discussed previously.
Basically, I added a BuildActorComponents() method to the GameActor that is called during the initialization of a GameActorProxy. This allows you to setup your actor components and initialize them without having to put them inside of BuildPropertyMaps(). Also, the BuildPropertyMap() method no longer adds the properties. Instead, the actor components add/remove the properties to the game actor when the actor component is added/removed. The BuildPropertyMap() of ActorComponents is now called from the Init() method. 

All of this was to solve a major problem: how do you initialize the values on your actor components before loading properties. It also solved a previous problem that was preventing you from adding/removing components except from within the CreateActor() function, which is a terrible place to add components.  It also addressed a big question of how & when the properties of an actor component should really get added to a Game Actor (actually, they shouldn't be added at all, but that is a task for another day).

Started the process of changing the dead reckoning helper into an actor component. This is only partially implemented and will need additional work.