Modified Game Actor Proxy to support accept deprecated properties in a ApplyactorUpdate() method.
Exposed the current instant velocity on the dead reckoning helper - passes it to the ground clamper. Added a clamp to the instant velocity for unusual cases like when you switch DR modes. Fixed a long standing bug with the ground clamper when there were 2 nearby surfaces above the dead reckoned location. Laid the ground work for a new property on the DR helper that will replace the 'Flying' property and allow vehicles to only ground clamp (upwards) if they go underground.

Did some minor cleanup in the dr helper to make the code easier to read.