Refactor the message system a bit.  Messages may now be auto registered by calling a different constructor on the message type class that takes the type.  Also added some macros to make creating a message type that uses the templated constructor easy and makes writing new message classes MUCH, MUCH easier.
many more files that would be expected are being committed because they were expecting message factory to include message type for them, which it now doesn't since message type has to call the message factory to auto register the message.  Plus, I had to move message classes from dtNetGM into dtGame because they had message TYPES in dtGame, but the classes were in dtNetGM, which was both weird and won't work with the autoregistration code.