This should actually be two commits, but they involve the same files, so they can't be disconnected and still produce working code.
The main commit is that the EnvironmentProcessRecordList list data type has been added to dtHLAGM to support EnvironmentProcess objects and their list of variant data.  Doing this involved some updates and changes to the named group parameter, which is now very capable and quite easy to use in code.  

The second commit is moving the CreateFromType static method back to named parameter.  It was moved off of named parameter at one point, but after a weekly call we decided to move it back.  That never happened apparently, so I did it while I was in there working on it.