Significant change to the way game actors work with networking and initial ownership.
Previously, the Initial Ownership flag was mostly ignored in the DefaultMessageProcessor. Now, it actually checks to see that you are flagged as a server or client, as appropriate, before adding the actor to the GM when loading a map. 

To do this, needed to add a way for the GM to have some status flag for whether it is performing the role of a server and/or client. Talked with David about this for a LONG time and decided that it would be bad to pigeon hole the GM with a single status enum. So, instead, added an extensible GMSettings class. This class currently has two pieces (IsServerRole and IsClientRole), but it can be subclassed to add custom behavior and you can set/get the GMSettings on the GM. 

By default, things work exactly as they did in the past. Meaning, the GMSettings assumes it's a server AND a client. Therefore, any initial ownership flag in a map will work like before. The MapChangeStateData will check the GMSettings and compare that with the Initial Ownership before deciding to add an actor from a map to the GM. So, you could have a 'Server_Local' actor on the server, but it wouldn't get added to the client. Sweet! But, the best case is having 'Server+Published' actors on the server that don't get added to clients until they get published! Booya!

To solve issues with partial updates caused by the new behavior (ie, a client joins late and gets rotation, but no mesh), the DefaultMessageProcessor will now ignore partial updates until an actor has been created (usually via a actor created or full actor update). The ActorUpdateMessage has a new param for IsPartialUpdate that defaults to false. On Game Actor Proxy, the NotifyPartialActorUpdate() methods allow an optional param to make this easy to use. Or, can set it manually on a new actor update message.

Everything is well commented and unit tested. Most users shouldn't notice anything different about their apps. But, apps that use networking will now have a TON more functionality. Simple to add, but huge new features for creating actors for the server or clients and putting them in your maps.