This big commit is a combination of both some name refactoring based on some future direction of delta3d, i.e., it renames ActorProxy to BaseActorObject and renames/adds options for many of the methods that have "proxy" in the name.  Many of the calls have been changed to use the newer methods names.
It also adds the PropertyContainerActorProperty and an alternate implementation of the ArrayActorProperty, which is poorly named ArrayActorPropertyComplex.  This new ArrayActorProperty has a much better interface on the actor that gets rid of the on-actor stick index, moving it to the property, and removes the need to constantly copy all the data in the array.

The property serializer now supports recursion of property containers to support the load save, but creation of the property containers could use some more sophistication.

The messaging support for arrays and property containers has been updated and improved, but it could use a few more unit tests.

The TestGamePropertyActor, has a property container property with an internal array pointing to the same object.  This allows for infinite recursion in stage and in map loading to test highly nested data.