Some internal GM Changes.  It now removes deleted actors before TICK_END_OF_FRAME because it had a race condition for anything that started a thread during that message.  It also runs a loop removing actors and handling messages so that delete messages for actors deleted by other actors can all be handled before the frame ends.
Also, reworked the shutdown of the GM so that it deletes the actors first and sends all delete messages immediately so that the components can clean up.  This fixes some common crash on exit problems we had.  In order to make map unload work the same way shutdown does, it now sends all delete messages when unloading map.  It doesn't send create messages on map load, be we have been looking at changing that too.