Extracted a pure-virtual interface for actorcomponentbase called ActorComponentContainer.
Now GameActorProxy implements that interface and so does actor component base.  Now you can call all actor component related methods on the proxy or the actor.  This will allow code to be migrated until the actorcomponentcontainer is pushed up to the ActorObject, which has yet to be extracted from GameActorProxy.