Committing a lot of changes to the project and xml handling code.  The basexml parser in dtDAL now loads through xerces with a stream.  This is a partial step to loading the xml based files via an osg reader writer.
The project now has a project config object, and a project config xml parser is added, but it's not all hooked up yet.  It will load via an reader writer and an istream/ostream so it can be zipped or put in an archive.  The map xml is also loaded via stream, and will be wrapped in a reader writer so it can be put in an archive and/or zipped.

Also reworked the xerces and constant init code so it doesn't require the project to know about all the parsers.  dtUtil also initialized xerces on startup now.

Also added zip as a valid extension to many of the resource types so they can be zipped in place and will still be selectable in stage.