+Adding a multipass mechansim with the new MultiPass Scene (brief below) +separating the cull masks out of the node masks +adding additional node masks for materials and lighting to be supported by shader composition (coming) +Adding shaders to the ephemeris as well as allowing for it to be on a separate pass +Modifications to SceneGroup to allow scene groups to cascade +SceneManager now has options for default scene or multipass mode also sets nodemasks on the appropriate scenes +revamping nodemask unit tests with cull mask tests +a small shader compiler error fix in multimap.frag
The new MultiPass Scene allows for a pre rendered depth and color pass as well as the ability to combine osgPPU effects.
This actor makes the main camera render to texture as well as adds 
an additional Multipass camera which pulls the main scene rendering into
a separate pass to allow multipass effects (reflections, refraction, etc) as well as post processing effects.