Added the lovely CivilianAIActorComponent.  It probably doesn't work yet.
Also changed the silly named RegisterForTicks() function to RegisterForTick() and made it intelligently decide whether to register for local or remote, and change the text label component not to have to do all the silliness associated with that.  I'm still not sure I'm sold on the API, but it at least doesn't sound stupid.