* Added MeshLampActor and LightActorComponent classes.
NOTE: LightActorComponent is used by MeshLampActor and will be used by FireworkActor. The component allows an actor to reference a DynamicLight and use its intensity to update a mesh's shader parameter for intensity. This should work for the lamp posts in TestApp once the relevant mesh actors change to the MeshLampActor (a subclass of MeshActorObject) and their shader parameter properties set.