* Added a fair bit of code to add shaders to the Particle Editor, for debugging shader effects that are to be used on particle systems. * Added code to prevent writing shaders to file. * Added code to reapply shaders after file write is completed. * Added controls for applying shaders to individual layers and added a button for recompiling the shaders.
NOTE: Some of the code was borrowed from the Animation Viewer for the same purpose. Once tools share common base functionality, the code can be moved to a place where the code can be shared.

TODO: Save the shader group name strings to some node in the particle system hierarchy.