* Added code to the PhysicsActComp for handling more than one PhysicsObject in an array property. * Added code to PhysicsObject for enabling/disabling the old property naming functionality. * Added associated unit tests. * Modified the PhysicsCompilerTool to take input geometry and to assign output physics mesh files to related PhysicsObjects that are generated. * Added a variable to VertexData for holding onto a resource descriptor, if a file is generated for it.
NOTE: PhysicsObject::RemoveOldProperties needs to be called at some point after map load in Stage so that old property names for the single PhysicsObject reference are removed prior to saving out a map.

NOTE: PhysicsObject array property needs to be able to save all properties to map when individual objects are modified.