A few more unit tests are needed, but removed the old properties from the physics object, and they just use the deprecated property system now to set the properties when a map is loaded.  Also fixed a compile error where the physics  actor component was trying to call the copy constructor on the physics object, which doesn't exist.
Also, the new insert and remove functions were not calling setUserData.  The adds and inserts now also call build property map, which has a guard call to make sure it doesn't run more than once.   This guard just checks to see if it has properties, which would be an issue of code tried to add properties before Build Property map is called, so this should probably be revisited at some point.

Also fixed some issues with indices not being the same sign-ness, which gave quite a few warnings.