* For now, adding the initial cut of new property controls to be used in the revamped property panel for Stage.
NOTE: These controls are still a work in progress. The controls may be abstracted away from properties to work with raw data without directly modifying data, for the sake of the undo system and interfacing with classes that do not have property objects. This is especially important for continuous update type of controls, such as sliders, which may change a value every frame.

An update in the near future will and an experimental new property panel to Stage that utilizes the new controls. Some control types still have to be implemented for the more complex or least common property data types.