Undoing in nested property containers and arrays works much better than it did.  Undoing array elements does not yet seem to work.
It uses updated functionalitiy in the named parameter code, which now handles partial change sets much better.  

The named array can be set to only modify certain elements, and add/remove elements.

The properties for the array and the property container delegate their to/from string and data stream to the named parameter so the complex code doesn't have to be implemented twice.