This is difficult commit to split up.
static mesh loading has been substantially unified, though not completely,
in the object class, which is likely be renamed at some point, as well as
the "dtCore::Model" class.  

static mesh actor is dtActors has been removed beause GameMeshActor now uses
an Object directly and has the same properties.   The terrain mesh actor now
extends GameMeshActor, and only to add a terrain version of the resource
property that sets the same internal variables.

Taking advantage of the dynamic actor components addition in the maps, a
considerable amount of code was deleted from the testApp.  

The changes in the GameMeshActor and StaticMeshActor required a number of
changes in unit tests, examples, and the fire fighter demo.  I'm not sure
yet if the firefight demo works.

The GameManager now supports lambda functions, but this also means that
passing a functor by reference doesn't work, which broke a unit test, which
has been changed to use lambda function.

The object simply needs to be changed to load meshes on a background thread
with relevant signals and slots, and all the mesh loading in simcore can be
replaced with this code.

Moved the sailboat into the vehicles subdirectory in examples/data/StaticMeshes
because the attach code is using "vehicle" a key to know if it can attach to
something.  Before it used the actor type as well as that, but the separate
actor types no longer exist.