CH-22
Modifications to fix the subsequent scenario slowdown, this was somehow caused by a realloc from an array reserve, which was only a problem after some reorganization of code, which is very peculiar.

Also, some improvements to the snow reset, there were cases where either empty nodes could be left around and it was not spatially optimized.